[cd] - Modify GA4 Custom Definitions
This change adds the ability to archive or create either a custom
dimension or custom metric in a GA4 property. These functions use the
same sheet as the list function, so all GA4 custom dimension or custom
metric use cases are contained in two sheets.

This change also modifies how GA4 account summaries are listed to
account for multiple page responses. Now, all account summaries are
properly written to the sheet.

Change-Id: If7b4a0ae4dbec081840e24d0e899e653d03e2be7