[error handling] - Improved Error Handling
Fixed some bugs that caused actions to be recorded as "created" or "updated" when an error occurred. The errors will now be properly
record in the corresponding Google Sheet. Whenever a GA4 setting is
listed, its sheet will be cleared of values before the new information is
written to the sheet.

Change-Id: Ibd3d1524edf2d603d8b8814e74012d24b5228bdf