[cloud] - Update Cloud Function
Starting to break the existing cloud function into different pieces so that different functions are invoked when specific resources need to be
created. This should reduce the complexity of each function and give
each function a discrete task it must perform.

This function creates ads links in Google Analytics based on the values
pushed in a pub/sub topic.

Change-Id: Ia179e0008544be491571c85709b06b9c1e5e02f2