[cloud] - Ads Links
I rewrote the ads_links function to be more generic. It can
now create, update, delete, and list Ads Links in Google Analytics.
It also properly decodes pubsub message data. This will make it easier
to deploy the functions and reduce the number of functions and pubsub
topics that will be needed to cover all the Admin API resources.

The function will perform a specific action depending on the action
value in the pubsub message.

Change-Id: I6e67c4f8143588440563e88f316009556a871337